User Management, Scheme Officials, Division and Scheme Form validation and re-designing

+++ src/app/pages/jbpm/divisions/division.component.ts
@@ -1,10 +1,11 @@
 import { Component, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
-import { NbStepperComponent } from '@nebular/theme';
+import { NbDialogService, NbGlobalPhysicalPosition, NbStepperComponent, NbToastrService } from '@nebular/theme';
 import {LocalDataSource} from 'ng2-smart-table';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {HttpClient} from '@angular/common/http';
 import { DepartmentManagementService } from '../../../jbpm/service/department-management.service';
 import {division_management_table_settings} from './divisions-utils';
+import { ConfirmDialogComponent } from '../../../jbpm/common-component/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'ngx-division',
@@ -26,8 +27,12 @@ export class DivisionComponent implements OnInit {
 
   readonly root: string = 'divisions';
   label: string = 'Add';
+  submitted = false;
+  loading =false;
 
   constructor(private formBuilder: FormBuilder,
+    private dialogService: NbDialogService,
+    private toastrService: NbToastrService,
               protected http: HttpClient, private service: DepartmentManagementService) {
     this.loadData();
   }
@@ -67,39 +72,84 @@ export class DivisionComponent implements OnInit {
   }
 
   departmentformSubmit(form: FormGroup): void {
-
+    this.submitted = true;
+    if (this.departmentForm.invalid) {
+      return;
+    }
     const  departmentBody: {[ k: string]: any} = {
       name  : form.value.name,
       description : form.value.description,
     };
-
+    this.loading = true;
     if (form.value?.departmentObject === null || form.value?.departmentObject.length === 0) {
       this.service.addDepartment(departmentBody, this.root).subscribe( res => {
+        this.toastrService.show(
+          'Divisions succesfully saved',
+          `Divisions has been succesfully saved`,
+          { 'position': NbGlobalPhysicalPosition.TOP_RIGHT, 'status': 'success' });
+          this.loading =false;
         this.source.prepend(res);
         this.onReset();
+      }, error => {
+        this.loading = false;
+        this.handleError();
       });
 
     } else if ( form.value?.departmentObject?._links.self?.href) {
       console.error(form.value?.departmentObject?._links.self?.href.replace());
       this.service.updateDepartment(form.value?.departmentObject?._links.self?.href, departmentBody )
         .subscribe( res => {
+          this.toastrService.show(
+            'Divisions succesfully saved',
+            `Divisions has been succesfully saved`,
+            { 'position': NbGlobalPhysicalPosition.TOP_RIGHT, 'status': 'success' });
+            this.loading =false;
           this.source.remove(form.value?.departmentObject).then(value => this.source.prepend(res));
           this.onReset();
+        }, error => {
+          this.loading = false;
+          this.handleError();
         });
     }
   }
 
-  deleteDepartment(form: FormGroup): void {
+  async deleteDepartment(form: FormGroup) {
+    this.loading = true;
+    const result = await this.dialogService.open(ConfirmDialogComponent, {
+      hasBackdrop: true, context: {
+        title: 'Delete Divisions',
+        message: 'Are you sure you want to delete?',
+        confirmText: 'Yes',
+        cancelText: 'No'
+      }
+    }).onClose.toPromise();
+    if (result) {
     this.service.deleteDepartment( form.value?.departmentObject?._links.self?.href)
       .subscribe(value_ => {
+        this.toastrService.show(
+          'Divisions removed',
+          `Divisions has been remove`,
+          { 'position': NbGlobalPhysicalPosition.TOP_RIGHT, 'status': 'success' });
+          this.loading =false;
         this.source.remove(form.value?.departmentObject).then(value => console.error(value) );
         this.onReset();
       });
+    }else{
+      this.loading =false;
+    }
+
   }
 
   onReset() {
     this.label = 'Add';
+    this.submitted=false;
     this.departmentForm.reset();
   }
-
+  get form() { return this.departmentForm.controls; }
+  handleError() {
+    this.toastrService.show(
+      'Error',
+      `Something went wrong`,
+      { 'position': NbGlobalPhysicalPosition.TOP_RIGHT, 'status': 'danger' });
+  }
 }
